% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_discrete.R
\name{any_discrete}
\alias{any_discrete}
\title{Any discrete variables}
\usage{
any_discrete(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Function checks if there are any discrete variables in a set datafame with op
}
