% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_numeric.R
\name{any_numeric}
\alias{any_numeric}
\title{Any numeric variables}
\usage{
any_numeric(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Function checks if there are any numeric variables in a set datafame with op
}
