% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.2.eksp.R
\name{chi.2.eksp}
\alias{chi.2.eksp}
\title{Chi Squared Experiment}
\usage{
chi.2.eksp(data = des, out.idx = grep("piqola", names(data)),
  min.length = 4 * 16, max.levels = 12, ver.idx = NULL, out.raw = FALSE,
  out.raw.name = "out.raw", out.df = TRUE, out.df.name = "p.df",
  out.list = TRUE, out.list.name = "spdf", p.cutoff = 0.05, dig = 3,
  suiwer = TRUE, verbose = TRUE, debug = TRUE, xlsx = TRUE,
  outdir = datadir, xlsx.name = "PIQOLA_chi2_uitkomste.xlsx", plot = TRUE,
  net.name = paste(naam, "chinet"))
}
\arguments{
\item{data}{an R object (data set)}

\item{out.idx}{Integer referring to the index of the name that matches "piqola"}

\item{min.length}{Numeric referring to the minimum length}

\item{max.levels}{Numeric referring to the maximum amount of levels}

\item{ver.idx}{Integer used in forecasting}

\item{out.raw}{Logical to display raw list}

\item{out.raw.name}{Character vector containing the name of the raw list}

\item{out.df}{Logical to display data frame}

\item{out.df.name}{Character vector containing the name of the data frame}

\item{out.list}{Logical to display list}

\item{out.list.name}{Character vector containing the name of the list}

\item{p.cutoff}{Numeric that refers to the probability coëfficient}

\item{dig}{Numeric that refers to the number of digits}

\item{suiwer}{Logical to leave those that give warnings}

\item{verbose}{Logical to display messages}

\item{debug}{Logical for debugging process}

\item{xlsx}{Logical to write xlsx file}

\item{outdir}{Character vector that contains the output directory}

\item{xlsx.name}{Character vector containing the name of the written xlsx file}

\item{plot}{Logical whether to plot the data}

\item{net.name}{Character vector containing the name of the gml network diagram}
}
\description{
Chi.2.eks with custom chi2 function that generates a chi2 statistic as well as a craemer's v
for a given input. The output is displayed in terms of outputs and exposures in the form of the data frame p.df
and list spdf
}

