% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{discrete_table}
\alias{discrete_table}
\title{Discrete Table}
\usage{
discrete_table(dvars, maxlevels = 20, ..., verbose = FALSE,
  forcegvar = FALSE, varSizeN = "0.15", levSizeN = "0.05")
}
\arguments{
\item{dvars}{Data frame containing the variables to be tested}

\item{maxlevels}{Numeric containing the maximum amount of levels}

\item{...}{Arguments passed down from the calling function}

\item{verbose}{Logical to display function messages}
}
\description{
This function uses the any_discrete and only_discrete functions to isolate discrete variables, 
whereby it uses the custom tableNominal2 function to create a LaTeX table of descriptive statistics
}

