<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table_write.R
\name{discrete_table_write}
\alias{discrete_table_write}
\title{Write Discrete Table}
\usage{
discrete_table_write(dvars, group = 1, d.cap = "", d.lab = "",
  maxlevels = 10, fn)
}
\arguments{
\item{dvars}{Data frame containing the nominal variables}

\item{group}{Optional grouping vector}

\item{d.cap}{The caption of the resulting LaTeX table}

\item{d.lab}{The lable of the resulting LaTeX table}

\item{maxlevels}{Numerical containing the maximum amount of levels}

\item{fn}{Character vector containing the desired name of the output file}
}
\description{
This function is used to write the discrete table to a file in the working directory. This
file contains the source code for a LaTeX document
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table_write.R
\name{discrete_table_write}
\alias{discrete_table_write}
\title{Write Discrete Table}
\usage{
discrete_table_write(dvars, group = 1, d.cap = "", d.lab = "",
  maxlevels = 10, fn)
}
\arguments{
\item{dvars}{Data frame containing the nominal variables}

\item{group}{Optional grouping vector}

\item{d.cap}{The caption of the resulting LaTeX table}

\item{d.lab}{The lable of the resulting LaTeX table}

\item{maxlevels}{Numerical containing the maximum amount of levels}

\item{fn}{Character vector containing the desired name of the output file}
}
\description{
This function is used to write the discrete table to a file in the working directory. This
file contains the source code for a LaTeX document
}
>>>>>>> dcbba8f7b4b45bbfe78c3fc073d535ca49fc6cad
