% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTextConvenience.R
\name{getalenkelmeer}
\alias{getalenkelmeer}
\title{getalenkelmeer}
\usage{
getalenkelmeer(vr = "stand_structures_number_formal",
  enkel = "structure", meer = NULL, df = dt, ia = FALSE)
}
\arguments{
\item{vr}{Charackter}

\item{enkel}{Charackter}

\item{meer}{Logical}

\item{df}{Dataframe}

\item{ia}{Logical}
}
\description{
funksie wat getalnaam, en enkel/meervoude kombineer
}
