% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTextConvenience.R
\name{getalnaam}
\alias{getalnaam}
\title{getalnaam}
\usage{
getalnaam(x, upperc = FALSE)
}
\arguments{
\item{x}{forseerbaar tot heelgetal}

\item{upperc}{Logies. Indien waar, sal woordresultaat se eerste letter 'n hoofletter gemaak word.}
}
\description{
funksie om getalle een tot tien in woorde om te sit
}
