% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.confounder.net.R
\name{kry.uitkoms.df}
\alias{kry.uitkoms.df}
\title{HELPER FUNCTION}
\usage{
kry.uitkoms.df(varname = "town", outname = "outcome",
  expname = "exposure", df, verbose = TRUE)
}
\arguments{
\item{varname}{Character vector}

\item{expname}{Character vector}

\item{df}{Data frame as used in chi.confounder.net}

\item{verbose}{**verbose not used}

\item{outcome}{Character vector}
}
\description{
Helper function for chi.confounder.net
}

