% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.opsie.tabel.R
\name{multi.opsie.tabel}
\alias{multi.opsie.tabel}
\title{Turn data into a multi option table
Finds the data related to yes and no categories}
\usage{
multi.opsie.tabel(df = des, x = "body_disease_diagnosed_month",
  yescat = "Yes", nocat = "No", tabeleer = TRUE, abs.perc = FALSE,
  tex = TRUE, tabdir = "~/", graphdir = "~/", fn = paste(x,
  "h.bar.", sep = ""), verbose = FALSE, fplace = "!ht")
}
\arguments{
\item{df}{The data frame that is used by the fuction e.g EIS dataset}

\item{x}{Character vector. For the selected category within the data to be used, e.g ceiling problems}

\item{yescat}{Character vector indicating "yes" category}

\item{nocat}{Character vector indicating "no" category}

\item{tabeleer}{Logical whether to create LaTeX table or not}

\item{abs.perc}{Logical that includes percentage of total results information when TRUE}

\item{tex}{Logical that prints LaTeX table when TRUE}

\item{tabdir}{Directory in which tables will be stored as Character vector}

\item{graphdir}{Directory in which graphs will be stored as Character vector}

\item{fn}{File name to be attributed to saved tables as Character vector}

\item{verbose}{Helps the user in debugging by sending messages in the console}
}
\description{
Turn data into a multi option table
Finds the data related to yes and no categories
}
