% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_report.R
\name{multi_report}
\alias{multi_report}
\title{Multi-report}
\usage{
multi_report(x, single = TRUE, vargroep_lbl = "een", var = NULL,
  groep = NULL, groeplab = NULL, out = TRUE, stats = c("all",
  "basic")[1], longt = FALSE, qs, qs_name = "question.name",
  qs_text = "question.text", drr = "~/tmp/", ncat = 4, nnum = 10,
  druk = FALSE, webify = FALSE, verbose = FALSE, wd = 15,
  md = TRUE, chi.digits = 4, debug = TRUE, varSizeN = "0.25",
  levSizeN = "0.15", voorspellende_veranderlikes = NULL, ...)
}
\description{
Multi-report
}
