% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.2.eksp.R
\name{my.chisq.test}
\alias{my.chisq.test}
\title{HELPER FUNCTION}
\usage{
my.chisq.test(x, y = NULL, correct = TRUE, p = rep(1/length(x),
  length(x)), rescale.p = FALSE, simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{x}{Either a data frame or a matrix}

\item{y}{R object that x will be compared to}

\item{correct}{Logical that applies Yates continuity correction}

\item{p}{Integer that applies the replicate function over the length of x}

\item{rescale.p}{Logical to rescale the probability}

\item{simulate.p.value}{Logical that simulates a probability value}

\item{Numeric}{that describes the number of replicates the simulated p value is based on}
}
\description{
Helper function called my.chisq.test that also supplies n and craemer's v. Test this vs 
assocstats in vcd package
}

