% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_table.R
\name{numeric_table_write}
\alias{numeric_table_write}
\title{Numeric Table Write}
\usage{
numeric_table_write(nvarss, fn, ...)
}
\arguments{
\item{nvarss}{Data frame containing continuous variables}

\item{fn}{A file name or a connection, if NULL returns the output as a character vector}

\item{...}{Arguments passed down from the calling function}
}
\description{
Writes the numeric table to a file in the chosen directory
}

