<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_discrete.R
\name{only_discrete}
\alias{only_discrete}
\title{Isolate discrete varables}
\usage{
only_discrete(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Function isolates only the discrete variables (use x <- only_discrete(x))
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_discrete.R
\name{only_discrete}
\alias{only_discrete}
\title{Isolate discrete varables}
\usage{
only_discrete(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Function isolates only the discrete variables (use x <- only_discrete(x))
}
>>>>>>> dcbba8f7b4b45bbfe78c3fc073d535ca49fc6cad
