<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_numeric.R
\name{only_numeric}
\alias{only_numeric}
\title{Isolate numeric varables}
\usage{
only_numeric(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Function isolates only the numeric variables (use x <- only_numeric(x))
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_numeric.R
\name{only_numeric}
\alias{only_numeric}
\title{Isolate numeric varables}
\usage{
only_numeric(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Function isolates only the numeric variables (use x <- only_numeric(x))
}
>>>>>>> dcbba8f7b4b45bbfe78c3fc073d535ca49fc6cad
