% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTextConvenience.R
\name{praat}
\alias{praat}
\title{praat}
\usage{
praat(vr = "id_household", d = dt, plak = TRUE, digits = 6,
  verbose = FALSE, hoofklein = TRUE, klein = FALSE)
}
\arguments{
\item{vr}{karakter. naam van 'n kolom in d}

\item{d}{dataraam}

\item{plak}{logies. Moet goed bymekaargesit word}

\item{digits}{heelgetal. Hoeveel syfers na desimaal}

\item{verbose}{logies Praat terug of nie}
}
\description{
geriefsfunksie om tbl waarde binne-in Rmd te gebruik

neem waarde van var uit d en druk dit
}
