% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.text.R
\name{qs.text}
\alias{qs.text}
\title{Text Output}
\usage{
qs.text(qs, varname = "question.name", question.text = "question.text")
}
\arguments{
\item{qs}{Questions data frame}

\item{varname}{Character vector referring to the name of the question}

\item{question.text}{Character string referring to the text of the question}
}
\description{
This function creates a text output from questions and returns 
a vector containing the question name and its text grouped together
}

