% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_report2.R
\name{section_report2}
\alias{section_report2}
\title{section report II}
\usage{
section_report2(x, qss = qs, sec = "A", sectionvarname = "section",
  question.text = "question.text", qs.varname = "question.name",
  groupvar = "", multicol = "question.type", multiname = "M",
  debug = FALSE)
}
\description{
section report II
}

