<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_woorde.R
\name{t_test_woorde}
\alias{t_test_woorde}
\title{t_test_woorde}
\usage{
t_test_woorde(df, vr, gp, alpha = 0.05, short = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{vr}{numeric Variable in terms of which T-test is performed}

\item{gp}{character or factor Name of grouping variable. Will be coerced to a factor with two levels}
}
\description{
Performs and reports on a T-test
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_woorde.R
\name{t_test_woorde}
\alias{t_test_woorde}
\title{t_test_woorde}
\usage{
t_test_woorde(df, vr, gp, alpha = 0.05, short = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{vr}{numeric Variable in terms of which T-test is performed}

\item{gp}{character or factor Name of grouping variable. Will be coerced to a factor with two levels}
}
\description{
Performs and reports on a T-test
}
>>>>>>> dcbba8f7b4b45bbfe78c3fc073d535ca49fc6cad
