% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableNominal2.R
\name{tableNominal2}
\alias{tableNominal2}
\title{tableNominal2}
\usage{
tableNominal2(vars, weights = NA, subset = NA, group = NA,
  miss.cat = NA, print.pval = c("none", "fisher", "chi2"),
  pval.bound = 10^-4, fisher.B = 2000, vertical = TRUE, cap = "",
  lab = "", col.tit.font = c("bf", "", "sf", "it", "rm"),
  font.size = "footnotesize", longtable = TRUE, nams = NA,
  cumsum = FALSE, debug = FALSE, ...)
}
\arguments{
\item{vars}{A list of nominal variables}

\item{weights}{Optional vector of weights of each observation}

\item{subset}{Optional logical vector, indicates subset of observations to be used}

\item{group}{Optional grouping vector}

\item{miss.cat}{Vector specifying the factors in vars that
should have their NAs transformed to a separate category}

\item{print.pval}{Add p-values of Fisher's exact or χ^2 test for a difference of distributions
between groups to the table, if there is more than one group}

\item{pval.bound}{p-values below pval.bound are formatted as < pval.bound}

\item{fisher.B}{Number of simulations to compute p-value for Fisher's exact test. Note that in the function fisher.test the option simulate.p.value
is set to TRUE, unless fisher.B == Inf which asks for the exact computation}

\item{vertical}{If TRUE, add vertical lines to the table, separating labels and groups, if applicable}

\item{cap}{The caption of the resulting LaTeX table}

\item{lab}{The label of the resulting LaTeX table}

\item{col.tit.font}{Choose the font for the column titles here (default: boldface)}

\item{font.size}{Font size for the generated table in LaTeX}

\item{longtable}{If TRUE, function makes use of package longtable in LaTex to generate tables that
span more than one page. If FALSE, generates a table in tabular environment}

\item{nams}{A vector of strings, containing the names corresponding to the variables in vars, if vars is
not a data frame but a list of variables. These are then the names that
appear in the LaTeX table. This option is only kept for backward compatibility.}

\item{cumsum}{f TRUE, the cumulative sums of the percentages are included for every level of the grouping variable}

\item{...}{Arguments pass through to print.xtable}

\item{varSizeN}{Size of the column containing the variables}

\item{levSizeN}{Size of the column containing the levels}
}
\description{
This function takes a data frame of nominal variables and possible grouping, weighting and subset
variables and provides a LaTeX table of descriptive statistics seperately per group
and jointly for all observations, per variable
}
