<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withWarnings.R
\name{withWarnings}
\alias{withWarnings}
\title{With Warnings}
\usage{
withWarnings(expr)
}
\arguments{
\item{expr}{expression to be evaluated}
}
\description{
This function is used by chi.2.eksp to ensure that the function works correctly. It makes use
of functions that provide a mechanism for handling unusual conditions, including errors and warnings. 
A list is returned of the value and the associated warning
}
\references{
Bill Dunlap
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withWarnings.R
\name{withWarnings}
\alias{withWarnings}
\title{With Warnings}
\usage{
withWarnings(expr)
}
\arguments{
\item{expr}{expression to be evaluated}
}
\description{
This function is used by chi.2.eksp to ensure that the function works correctly. It makes use
of functions that provide a mechanism for handling unusual conditions, including errors and warnings. 
A list is returned of the value and the associated warning
}
\references{
Bill Dunlap
}
>>>>>>> dcbba8f7b4b45bbfe78c3fc073d535ca49fc6cad
