% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTextConvenience.R
\name{woodcoal}
\alias{woodcoal}
\title{woodcoal}
\usage{
woodcoal(
  df = dt,
  coal = "solid_fuel_use_coal",
  wood = "solid_fuel_use_wood",
  yesval = "yes"
)
}
\arguments{
\item{df}{dataframe}

\item{coal}{Character}

\item{wood}{Character}
}
\description{
woodcoal
}
