% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afh.tab.R
\name{afh.tab}
\alias{afh.tab}
\title{Dependancy tables}
\usage{
afh.tab(x, alpha = 0.05, outcome = "outcome",
  var.names = "var.names", exposure = "exposure",
  p.value = "p.value", craemer = "craemer", df = "df", n = "n",
  statistic = "statistic", x.names = c("Related variable", "n",
  "p value", "df", paste("$\\\\mathbf{\\\\chi^2}$", "statistic"),
  "Craemer's V"), verkort = FALSE, warnings = "warnings",
  sonder.warnings = TRUE, geen = "Geen", tp = "H", tabdir = NULL,
  verbose = FALSE, bonferroni = TRUE)
}
\arguments{
\item{x}{Data frame, which is the outcome of chi.2.exp}

\item{alpha}{Numeric to be used in Bonferroni}

\item{outcome}{Character vector referring to the column containing outcomes}

\item{exposure}{Character vector referring to the column containing exposures}

\item{p.value}{Character vector referring to the column containing the probability value}

\item{craemer}{Character vector referring to the column containing creamer values}

\item{df}{Character vector referring to the column containing degrees of freedom}

\item{n}{Character vector referring to the column containing the total number of observations}

\item{statistic}{Character vector referring to the column containing the
calculated chi2 statistic}

\item{warnings}{Character referring to the column containing warnings}

\item{sonder.warnings}{Logical referring to no warnings being present if TRUE}

\item{geen}{Character vector to indicating that no warnings are present}

\item{tp}{Character vector for the table placement}

\item{tabdir}{Directory of the table}

\item{verbose}{Display function messages}

\item{bonferroni}{Logical that makes a bonferroni adjustment when TRUE}

\item{lateX}{Logical that prints a LaTeX table if TRUE}

\item{markDown}{Logical that creates a table for use in MarkDown}

\item{latSave}{Logical that saves the tables in the Tabelle folder in your working directory}
}
\description{
Prints out tables for report out of data frame p.df, which is the result of the chi.2.exp function.
Bonferroni's method is used here, which is used to correct for multiple comparisons
}
