% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.confounder.net.R
\name{chi.confounder.net}
\alias{chi.confounder.net}
\title{Confounder Chi-net}
\usage{
chi.confounder.net(df, outcome = "outcome", exposure = "exposure",
  outdir = "graphs/", stoor = TRUE, remPiq = FALSE)
}
\arguments{
\item{df}{Data frame containing outcomes and exposures}

\item{outcome}{Character vector referring to the name of the outcome column}

\item{exposure}{Character vector referring to the name of the exposure column}

\item{outdir}{Character vector containing to output directory}

\item{stoor}{Logical that stores the file as pdf when TRUE}

\item{Logical}{that removes "Piqola_" from outcomes and exposures when TRUE}
}
\description{
Function to graphically display chi2 test results to 
ID confounders. Requires a data frame with outcome and exposures 
such as the one provided by the chi.2.eksp function. Returns a list of network objects in pdf
}

