% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTextConvenience.R
\name{getalnaam}
\alias{getalnaam}
\title{getalnaam}
\usage{
getalnaam(x)
}
\arguments{
\item{x}{forseerbaar tot heelgetal}
}
\description{
funksie om getalle een tot tien in woorde om te sit
}
