<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.confounder.net.R
\name{kry.uitkoms.df}
\alias{kry.uitkoms.df}
\title{kry.uitkoms.df}
\usage{
kry.uitkoms.df(varname = "town", outname = "outcome",
  expname = "exposure", df, remPiq = FALSE)
}
\arguments{
\item{varname}{Character vector of the variable name}

\item{outname}{Character vector referring to the name of the outcome column}

\item{expname}{Character vector referring to the name of the exposure column}

\item{df}{Data frame as used in chi.confounder.net}

\item{remPiq}{Logical that removes "Piqola_" from outcomes and exposures when TRUE}
}
\description{
Helper function for chi.confounder.net
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.confounder.net.R
\name{kry.uitkoms.df}
\alias{kry.uitkoms.df}
\title{Kry Uitkoms}
\usage{
kry.uitkoms.df(varname = "town", outname = "outcome",
  expname = "exposure", df, remPiq = FALSE)
}
\arguments{
\item{varname}{Character vector of the variable name}

\item{outname}{Character vector referring to the name of the outcome column}

\item{expname}{Character vector referring to the name of the exposure column}

\item{df}{Data frame as used in chi.confounder.net}

\item{remPiq}{Logical that removes "Piqola_" from outcomes and exposures when TRUE}
}
\description{
Helper function for chi.confounder.net
}
>>>>>>> dcbba8f7b4b45bbfe78c3fc073d535ca49fc6cad
