% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_table.R
\name{multi_table}
\alias{multi_table}
\title{Multi Table}
\usage{
multi_table(vars = vars, qs = NULL, multicol = "question.type",
  multiname = "M", question.name = "question.name", maxlevels = 20,
  verbose = FALSE, ...)
}
\arguments{
\item{vars}{A data frame of nominal variables such as DES}

\item{qs}{Data frame containing question data}

\item{multicol}{Character vector containing the question type}

\item{multiname}{Character vector containing the name of multiple choice questions}

\item{question.name}{Character vector containing the name of the question}

\item{maxlevels}{Numeric containing the maximum amount of levels}

\item{verbose}{Logical to display messages}

\item{...}{Arguments passed down from the calling function}
}
\description{
This function is used to test for discrete variables using any_discrete and then isolates them using only_discrete.
It includes tests that identify if there are any multiple choice questions whereby it   
generates a table showing answer percentages for each variable
}
