% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_table.R
\name{numeric_table}
\alias{numeric_table}
\title{Numeric Table}
\usage{
numeric_table(nvars, verbose = FALSE, ...)
}
\arguments{
\item{nvars}{Data frame}

\item{verbose}{Logical to display messages}

\item{...}{Arguments passed down from the calling function}
}
\description{
This function
}

