<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.text.R
\name{qs.text}
\alias{qs.text}
\title{qs.text}
\usage{
qs.text(qs, varname = "question_name", question.text = "question_text")
}
\arguments{
\item{qs}{Questions data frame}

\item{varname}{Character vector referring to the question name column}

\item{question.text}{Character vector referring to the question text columnn}
}
\description{
This function creates a text output from questions and returns 
a vector containing the question name and its text grouped together
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.text.R
\name{qs.text}
\alias{qs.text}
\title{qs.text}
\usage{
qs.text(qs, varname = "question_name", question.text = "question_text")
}
\arguments{
\item{qs}{Questions data frame}

\item{varname}{Character vector referring to the question name column}

\item{question.text}{Character vector referring to the question text columnn}
}
\description{
This function creates a text output from questions and returns 
a vector containing the question name and its text grouped together
}
>>>>>>> dcbba8f7b4b45bbfe78c3fc073d535ca49fc6cad
