% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.text.R
\name{qs.text.string}
\alias{qs.text.string}
\title{String Output}
\usage{
qs.text.string(qs, varname = "question.name",
  question.text = "question.text")
}
\arguments{
\item{qs}{Questions data frame}

\item{varname}{Character vector referring to the name of the question}

\item{question.text}{Character string referring to the text of the question}
}
\description{
This function behaves the same as qs.text, but it manipulates a character vector of format strings
}

