% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_report2.R
\name{section_report2}
\alias{section_report2}
\title{section report II}
\usage{
section_report2(x, qss = qs, sec = "A", sectionvarname = "section",
  question.text = "question.text", qs.varname = "question.name",
  groupvar = "", multicol = "question.type", multiname = "M",
  debug = FALSE, forcegvar = FALSE, varSizeN = "0.15",
  levSizeN = "0.05", varSizeC = "0.15", levSizeC = "0.15")
}
\arguments{
\item{x}{Data frame containing the survey results, typically the DES}

\item{qss}{Data frame containing the questions and responses}

\item{sec}{Character vector referring to the section under consideration}

\item{sectionvarname}{Character vector containing the name of the section column}

\item{question.text}{Character vector containing the name of the question text column}

\item{qs.varname}{Character vector containing the question name column reference}

\item{groupvar}{Character vector containing the grouping variable under consideration}

\item{multicol}{Character vector containing the name of the question type column}

\item{multiname}{Character vector referring to the multiple choice questions}

\item{debug}{Logical to display function steps}

\item{forcegvar}{Logical that forces a large amount of grouping variables to be processed if TRUE. 
This will cause an ugly and unrefined table output}

\item{varSizeN}{Character vector that contains the percentage column size to be
attributed to the column containing variables. This is used to customize tables. 
Used in function discrete_table.}

\item{levSizeN}{Character vector that contains the percentage column size to be
attributed to the column containing levels. This is used to customize tables.
Used in function discrete_table.}

\item{varSizeC}{Character vector that contains the percentage column size to be
attributed to the column containing variables. This is used to customize tables.
Used in function numeric_table.}

\item{levSizeC}{Character vector that contains the percentage column size to be
attributed to the column containing levels. This is used to customize tables.
Used in function numeric_table.}
}
\description{
This function creates a report based on a specific section as well as a grouping variable. The numeric_table, 
discrete_table and multi_table functions are used to generate the LaTeX output
}
